Update README.mdFeedback
The Feedback Management System is a comprehensive solution designed to simplify the process of gathering and analyzing user feedback. This web application allows users to submit their feedback 

The responsive design ensures accessibility across various devices, making it easy for users to provide feedback anytime, anywhere. Built using React for the frontend and Express.js for the backend, this system combines modern technology with intuitive design to create an effective feedback management tool.